requirements.txt
Django==2.0.7
pytz==2018.5
djangorestframework==3.9.0
